Add command to show parse tree
